tests/wpt/web-platform-tests/content-security-policy/plugin-types/plugintypes-notype-data.html

<!DOCTYPE html>
<html>

<head>
  <meta http-equiv="Content-Security-Policy" content="plugin-types application/pdf;">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>

<body>
  <script>
    var t = async_test('Should not load the object because it does not have a declared type');
    window.addEventListener('securitypolicyviolation', t.step_func_done(function(e) {
      assert_equals(e.violatedDirective, "plugin-types");
      assert_equals(e.blockedURI, "");
    }));
  </script>

  <!-- Objects need to declare an explicit type -->
  <object data="data:application/x-shockwave-flash,asdf"></object>
</body>

</html>
