Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/content-security-policy/plugin-types/plugintypes-notype-data-expected.txt

CONSOLE MESSAGE: Refused to load data:application/x-shockwave-flash,asdf because its MIME type does not appear in the plugin-types directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load data:application/x-shockwave-flash,asdf because its MIME type does not appear in the plugin-types directive of the Content Security Policy.


Harness Error (TIMEOUT), message = null

NOTRUN Should not load the object because it does not have a declared type
